requirements.txt
Flask==2.2.1
google-cloud-storage==1.31.0
google-cloud-bigquery==3.3.1
requests==2.27.1
python-dateutil==2.8.2